cleans container compilation
cs
