+++ src/main/core/ClarolineCoreBundle.php
@@ -12,8 +12,7 @@
 namespace Claroline\CoreBundle;
 
 use Bazinga\Bundle\JsTranslationBundle\BazingaJsTranslationBundle;
-use Claroline\CoreBundle\DependencyInjection\Compiler\DoctrineEntityListenerPass;
-use Claroline\CoreBundle\DependencyInjection\Compiler\DynamicConfigPass;
+use Claroline\CoreBundle\DependencyInjection\Compiler\AnonymousAuthenticationPass;
 use Claroline\CoreBundle\DependencyInjection\Compiler\GeoipPass;
 use Claroline\CoreBundle\DependencyInjection\Compiler\MailingConfigPass;
 use Claroline\CoreBundle\DependencyInjection\Compiler\MessengerConfigPass;
@@ -45,7 +44,7 @@ class ClarolineCoreBundle extends DistributionPluginBundle implements Configurat
         parent::build($container);
 
         $container->addCompilerPass(new PlatformConfigPass());
-        $container->addCompilerPass(new DynamicConfigPass());
+        $container->addCompilerPass(new AnonymousAuthenticationPass());
         $container->addCompilerPass(new MailingConfigPass());
         $container->addCompilerPass(new SessionConfigPass());
         $container->addCompilerPass(new MessengerConfigPass());
