+++ src/main/core/DependencyInjection/Compiler/MailingConfigPass.php
@@ -13,12 +13,26 @@ namespace Claroline\CoreBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Component\Mailer\Transport\TransportInterface;
+use Symfony\Component\Mailer\Transport\Transports;
 
 class MailingConfigPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container)
     {
+        $transport = new Definition();
+        $transport->setClass(TransportInterface::class);
+        $transport->setFactory([
+            new Reference('Claroline\CoreBundle\Library\Mailing\TransportFactory'),
+            'getTransport',
+        ]);
+        $container->removeDefinition('mailer.default_transport');
+        $container->setDefinition('mailer.default_transport', $transport);
+        $container->removeDefinition('mailer.transports');
+        $container->register('mailer.transports', Transports::class)->addArgument([new Reference('mailer.default_transport')]);
+
         if (false === $container->hasDefinition('Claroline\CoreBundle\Library\Mailing\Mailer')) {
             return;
         }
